order-api/src/main/java/com/example/orderapi/service/ProductService.java
package com.example.orderapi.service;


import com.example.orderapi.domain.model.Product;
import com.example.orderapi.domain.product.AddProductForm;
import com.example.orderapi.domain.repository.ProductRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class ProductService {
    private final ProductRepository productRepository;

    // 상품 추가
    @Transactional
    public Product addProduct(Long sellerId, AddProductForm form){
        return productRepository.save(Product.of(sellerId, form));
    }

}
